Reorganization. Added NewRoasting and Roasings forms. Edited data types in EF.

+++ Espresso/Entity/CoffeeSort.cs
@@ -28,7 +28,7 @@ namespace Espresso.Entity
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public Nullable<int> LastPrice { get; set; }
+        public Nullable<double> LastPrice { get; set; }
         public bool Active { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
